добвлен класс Converter и созданы для него тесты ConverterTest

+++ src/main/java/ru/job4j/Converter.java
@@ -1,11 +1,11 @@
 package ru.job4j;
 
 public class Converter {
-    public static double rubleToEuro(double value) {
+    public static float rubleToEuro(float value) {
         return value / 90;
     }
 
-    public static double rubleToDollar(double value) {
+    public static float rubleToDollar(float value) {
         return value / 75;
     }
 }
